+++ QuanLyThuVien/QuanLyThuVien/UC_DauSach.xaml.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Data;
 using System.Data.SqlClient;
+using System.Diagnostics.Eventing.Reader;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -30,10 +31,10 @@ namespace QuanLyThuVien
         public UC_DauSach()
         {
             InitializeComponent();
-            string connectionString = @"Data Source=.\;Initial Catalog=QLTV;Integrated Security=True;"; 
+            string connectionString = @"Data Source=DESKTOP-AV6EQV4\SQLEXPRESS;Initial Catalog=QLTV_DB;User ID=sa;Password=123456;Pooling=False;Encrypt=True;TrustServerCertificate=True";
             sqlConnection = new SqlConnection(connectionString);
-            InitMaTheLoai();
-            InitMaTacGia();
+            InitTenTheLoai();
+            InitTenTacGia();
             InitMaDauSach();
             HienThiDanhSachDauSach();
         }
@@ -42,7 +43,7 @@ namespace QuanLyThuVien
         private void HienThiDanhSachDauSach()
         {
             sqlConnection.Open();
-            string query = "SELECT * FROM DAUSACH";
+            string query = "SELECT MaDauSach AS 'Mã đầu sách', TenDauSach AS 'Tên đầu sách', TenTheLoai AS 'Tên thể loại' FROM DAUSACH JOIN THELOAI ON DAUSACH.MaTheLoai = THELOAI.MaTheLoai";
             SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
             SqlDataAdapter da = new SqlDataAdapter(sqlCommand);
             DataTable dt = new DataTable();
@@ -57,20 +58,20 @@ namespace QuanLyThuVien
             DataRowView row_selected = gd.SelectedItem as DataRowView;
             if (row_selected != null)
             {
-                tblMaDauSach.Text = row_selected.Row["MaDauSach"].ToString();
-                txbTenDauSach.Text = row_selected.Row["TenDauSach"].ToString();
-                cbMaTheLoai.Text = row_selected.Row["MaTheLoai"].ToString();
+                tblMaDauSach.Text = row_selected.Row["Mã đầu sách"].ToString();
+                txbTenDauSach.Text = row_selected.Row["Tên đầu sách"].ToString();
+                cbTenTheLoai.Text = row_selected.Row["Tên thể loại"].ToString();
 
                 lbTacGia.Items.Clear();
 
                 sqlConnection.Open();
-                SqlCommand command = new SqlCommand("SELECT MaTacGia FROM CT_TACGIA WHERE MaDauSach = @MaDauSach", sqlConnection);
-                command.Parameters.AddWithValue("@MaDauSach", row_selected.Row["MaDauSach"].ToString());
+                SqlCommand command = new SqlCommand("SELECT TenTacGia FROM CT_TACGIA JOIN TACGIA ON CT_TACGIA.MaTacGia = TACGIA.MaTacGia WHERE MaDauSach = @MaDauSach", sqlConnection);
+                command.Parameters.AddWithValue("@MaDauSach", row_selected.Row["Mã đầu sách"].ToString());
                 SqlDataReader reader = command.ExecuteReader();
 
                 while (reader.Read())
                 {
-                    string value = reader["MaTacGia"].ToString();
+                    string value = reader["TenTacGia"].ToString();
                     // Thêm giá trị vào danh sách hiển thị trong ListBox
                     lbTacGia.Items.Add(value);
                 }
@@ -80,9 +81,9 @@ namespace QuanLyThuVien
             
         }
 
-        private void InitMaTheLoai()
+        private void InitTenTheLoai()
         {
-            string query = "SELECT MaTheLoai FROM THELOAI"; // Thay thế YourTableName bằng tên bảng của bạn
+            string query = "SELECT TenTheLoai FROM THELOAI"; // Thay thế YourTableName bằng tên bảng của bạn
 
             sqlConnection.Open();
             SqlCommand command = new SqlCommand(query, sqlConnection);
@@ -90,15 +91,15 @@ namespace QuanLyThuVien
 
             while (reader.Read())
             {
-                string companyName = reader["MaTheLoai"].ToString();
-                cbMaTheLoai.Items.Add(companyName);
+                string companyName = reader["TenTheLoai"].ToString();
+                cbTenTheLoai.Items.Add(companyName);
             }
             sqlConnection.Close();
         }
 
-        private void InitMaTacGia()
+        private void InitTenTacGia()
         {
-            string query = "SELECT MaTacGia FROM TACGIA"; // Thay thế YourTableName bằng tên bảng của bạn
+            string query = "SELECT TenTacGia FROM TACGIA"; // Thay thế YourTableName bằng tên bảng của bạn
 
             sqlConnection.Open();
             SqlCommand command = new SqlCommand(query, sqlConnection);
@@ -106,7 +107,7 @@ namespace QuanLyThuVien
 
             while (reader.Read())
             {
-                string companyName = reader["MaTacGia"].ToString();
+                string companyName = reader["TenTacGia"].ToString();
                 cbTacGia.Items.Add(companyName);
             }
             sqlConnection.Close();
@@ -135,7 +136,7 @@ namespace QuanLyThuVien
         {
             tblMaDauSach.Text = IncreasePrimaryKey();
             txbTenDauSach.Text = "";
-            cbMaTheLoai.SelectedIndex = -1;
+            cbTenTheLoai.SelectedIndex = -1;
             cbTacGia.SelectedIndex = -1;
             lbTacGia.Items.Clear();
         }
@@ -168,23 +169,52 @@ namespace QuanLyThuVien
         {
             try
             {
-                
                 sqlConnection.Open();
-                string query = "INSERT INTO DAUSACH (MaDauSach, TenDauSach, MaTheLoai) VALUES (@MaDauSach, @TenDauSach, @MaTheLoai)";
+                string query = "SELECT * FROM DAUSACH WHERE MaDauSach = @MaDauSach";
                 SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
                 sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                sqlCommand.Parameters.AddWithValue("@TenDauSach", txbTenDauSach.Text);
-                sqlCommand.Parameters.AddWithValue("@MaTheLoai", cbMaTheLoai.Text);
-                sqlCommand.ExecuteScalar();
 
-                for (int i = 0; i < lbTacGia.Items.Count; i++)
+                if (sqlCommand.ExecuteScalar() != null)
+                    MessageBox.Show("Đã tồn tại đầu sách");
+                else if (tblMaDauSach.Text == "")
+                    MessageBox.Show("Vui lòng chọn 'Thêm mới' để nhập thông tin");
+                else if (txbTenDauSach.Text == "")
+                    MessageBox.Show("Tên đầu sách không được để trống");
+                else if (cbTenTheLoai.Text == "")
+                    MessageBox.Show("Tên thể loại không được để trống");
+                else if (lbTacGia.Items.Count == 0)
+                    MessageBox.Show("Tác giả của sách không được để trống");                
+                else
                 {
-                    var item = lbTacGia.Items[i];
-                    query = "INSERT INTO CT_TACGIA (MaDauSach, MaTacGia) VALUES (@MaDauSach, @MaTacGia)";
+                    query = "SELECT MaTheLoai FROM THELOAI WHERE TenTheLoai = @TenTheLoai";
+                    sqlCommand = new SqlCommand(query, sqlConnection);
+                    sqlCommand.Parameters.AddWithValue("@TenTheLoai", cbTenTheLoai.Text);
+                    string maTheLoai = sqlCommand.ExecuteScalar().ToString();
+
+                    query = "INSERT INTO DAUSACH (MaDauSach, TenDauSach, MaTheLoai) VALUES (@MaDauSach, @TenDauSach, @MaTheLoai)";
                     sqlCommand = new SqlCommand(query, sqlConnection);
                     sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                    sqlCommand.Parameters.AddWithValue("@MaTacGia", item.ToString());
+                    sqlCommand.Parameters.AddWithValue("@TenDauSach", txbTenDauSach.Text);
+                    sqlCommand.Parameters.AddWithValue("@MaTheLoai", maTheLoai);
                     sqlCommand.ExecuteScalar();
+
+                    for (int i = 0; i < lbTacGia.Items.Count; i++)
+                    {
+                        var tenTacGia = lbTacGia.Items[i];
+
+                        query = "SELECT MaTacGia FROM TACGIA WHERE TenTacGia = @TenTacGia";
+                        sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@TenTacGia", tenTacGia.ToString());
+                        string maTacGia = sqlCommand.ExecuteScalar().ToString();
+
+                        query = "INSERT INTO CT_TACGIA (MaDauSach, MaTacGia) VALUES (@MaDauSach, @MaTacGia)";
+                        sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
+                        sqlCommand.Parameters.AddWithValue("@MaTacGia", maTacGia);
+                        sqlCommand.ExecuteScalar();
+                    }
+
+                    MessageBox.Show("Thêm thành công");
                 }
                 
             }
@@ -196,6 +226,7 @@ namespace QuanLyThuVien
             {
                 sqlConnection.Close();
                 HienThiDanhSachDauSach();
+
             }
         }
 
@@ -204,27 +235,59 @@ namespace QuanLyThuVien
             try
             {
                 sqlConnection.Open();
-                string query = "UPDATE DAUSACH SET TenDauSach = @TenDauSach, MaTheLoai = @MaTheLoai WHERE MaDauSach = @MaDauSach";
+                string query = "SELECT * FROM DAUSACH WHERE MaDauSach = @MaDauSach";
                 SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
                 sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                sqlCommand.Parameters.AddWithValue("@TenDauSach", txbTenDauSach.Text);
-                sqlCommand.Parameters.AddWithValue("@MaTheLoai", cbMaTheLoai.Text);
-                sqlCommand.ExecuteScalar();
 
-                query = "DELETE FROM CT_TACGIA WHERE MaDauSach = @MaDauSach";
-                sqlCommand = new SqlCommand(query, sqlConnection);
-                sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                sqlCommand.ExecuteScalar();
-
-                for (int i = 0; i < lbTacGia.Items.Count; i++)
+                if (sqlCommand.ExecuteScalar() == null)
+                    MessageBox.Show("Không tồn tại đầu sách");
+                else if (tblMaDauSach.Text == "")
+                    MessageBox.Show("Vui lòng chọn 'Thêm mới' để nhập thông tin");
+                else if (txbTenDauSach.Text == "")
+                    MessageBox.Show("Tên đầu sách không được để trống");
+                else if (cbTenTheLoai.Text == "")
+                    MessageBox.Show("Tên thể loại không được để trống");
+                else if (lbTacGia.Items.Count == 0)
+                    MessageBox.Show("Tác giả của sách không được để trống");
+                else
                 {
-                    var item = lbTacGia.Items[i];
-                    query = "INSERT INTO CT_TACGIA (MaDauSach, MaTacGia) VALUES (@MaDauSach, @MaTacGia)";
+
+                    query = "SELECT MaTheLoai FROM THELOAI WHERE TenTheLoai = @TenTheLoai";
+                    sqlCommand = new SqlCommand(query, sqlConnection);
+                    sqlCommand.Parameters.AddWithValue("@TenTheLoai", cbTenTheLoai.Text);
+                    string maTheLoai = sqlCommand.ExecuteScalar().ToString();
+
+                    query = "UPDATE DAUSACH SET TenDauSach = @TenDauSach, MaTheLoai = @MaTheLoai WHERE MaDauSach = @MaDauSach";
                     sqlCommand = new SqlCommand(query, sqlConnection);
                     sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                    sqlCommand.Parameters.AddWithValue("@MaTacGia", item.ToString());
+                    sqlCommand.Parameters.AddWithValue("@TenDauSach", txbTenDauSach.Text);
+                    sqlCommand.Parameters.AddWithValue("@MaTheLoai", maTheLoai);
                     sqlCommand.ExecuteScalar();
+
+                    query = "DELETE FROM CT_TACGIA WHERE MaDauSach = @MaDauSach";
+                    sqlCommand = new SqlCommand(query, sqlConnection);
+                    sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
+                    sqlCommand.ExecuteScalar();
+
+                    for (int i = 0; i < lbTacGia.Items.Count; i++)
+                    {
+                        var tenTacGia = lbTacGia.Items[i];
+
+                        query = "SELECT MaTacGia FROM TACGIA WHERE TenTacGia = @TenTacGia";
+                        sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@TenTacGia", tenTacGia.ToString());
+                        string maTacGia = sqlCommand.ExecuteScalar().ToString();
+
+                        query = "INSERT INTO CT_TACGIA (MaDauSach, MaTacGia) VALUES (@MaDauSach, @MaTacGia)";
+                        sqlCommand = new SqlCommand(query, sqlConnection);
+                        sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
+                        sqlCommand.Parameters.AddWithValue("@MaTacGia", maTacGia);
+                        sqlCommand.ExecuteScalar();
+                    }
+
+                    MessageBox.Show("Cập nhật thành công");
                 }
+                
 
             }
             catch (Exception ex)
@@ -243,15 +306,31 @@ namespace QuanLyThuVien
             try
             {
                 sqlConnection.Open();
-                string query = "DELETE FROM CT_TACGIA WHERE MaDauSach = @MaDauSach";
+                string query = "SELECT * FROM DAUSACH WHERE MaDauSach = @MaDauSach";
                 SqlCommand sqlCommand = new SqlCommand(query, sqlConnection);
                 sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                sqlCommand.ExecuteScalar();
+                if (sqlCommand.ExecuteScalar() == null)
+                    MessageBox.Show("Không tìm thấy chi tiết phiếu");
+                else
+                {
+                    query = "DELETE FROM CT_TACGIA WHERE MaDauSach = @MaDauSach";
+                    sqlCommand = new SqlCommand(query, sqlConnection);
+                    sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
+                    sqlCommand.ExecuteScalar();
 
-                query = "DELETE FROM DAUSACH WHERE MaDauSach = @MaDauSach";
-                sqlCommand = new SqlCommand(query, sqlConnection);
-                sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
-                sqlCommand.ExecuteScalar();
+                    query = "DELETE FROM DAUSACH WHERE MaDauSach = @MaDauSach";
+                    sqlCommand = new SqlCommand(query, sqlConnection);
+                    sqlCommand.Parameters.AddWithValue("@MaDauSach", tblMaDauSach.Text);
+                    sqlCommand.ExecuteScalar();
+                }
+
+                tblMaDauSach.Text = "";
+                txbTenDauSach.Text = "";
+                cbTenTheLoai.SelectedIndex = -1;
+                cbTacGia.SelectedIndex = -1;
+                lbTacGia.Items.Clear();
+                                   
+                
             }
             catch (Exception ex)
             {
